Mod report & emergency report functionaliity
Added a response to a user if they submit a mod report. If the report matches certain words, then the report will show extra text that allows the user to flag the message as urgent, which will ping off a message in a matrix chat room for push notification.

Also some small bug fixes/tweaks