V 0.7.1
Alters the way in which spam messages are handled now that Lemmy 0.19.10 removes PMs as part of the account ban process - no direct connection required to the database now.

Change how a spam message is detected to include spaces in the message. Suspicious multiple URLs are instead instead flagged to admins (WIP)